# Optimize allocations - shared cursor struct for Put/Get/Bucket