Each bucket operation creates a new cursor with []itemRef
moving it to the Tx structure allows to share it and avoid allocations.

BenchmarkAlloc1024Op: 1024 operations in one transaction:

```
benchmark                               old ns/op      new ns/op      delta
BenchmarkDBBatchAutomatic-16            3461377        3337566        -3.58%
BenchmarkDBBatchSingle-16               1981311812     1979383688     -0.10%
BenchmarkAlloc1024Op/Get-16             3927441        3772682        -3.94%
BenchmarkAlloc1024Op/PutNew-16          5325346        5190664        -2.53%
BenchmarkAlloc1024Op/PutReplace-16      5140605        5018904        -2.37%
BenchmarkAlloc1024Op/SharedTxGet-16     227576         108000         -52.54%
BenchmarkDBBatchManual10x100-16         21471994       21324567       -0.69%

benchmark                               old allocs     new allocs     delta
BenchmarkDBBatchAutomatic-16            10194          8184           -19.72%
BenchmarkDBBatchSingle-16               56811          54913          -3.34%
BenchmarkAlloc1024Op/Get-16             2110           62             -97.06%
BenchmarkAlloc1024Op/PutNew-16          3377           1328           -60.68%
BenchmarkAlloc1024Op/PutReplace-16      3344           1295           -61.27%
BenchmarkAlloc1024Op/SharedTxGet-16     2048           0              -100.00%
BenchmarkDBBatchManual10x100-16         5937           3981           -32.95%

benchmark                               old bytes     new bytes     delta
BenchmarkDBBatchAutomatic-16            451744        378593        -16.19%
BenchmarkDBBatchSingle-16               28234584      28218672      -0.06%
BenchmarkAlloc1024Op/Get-16             87860         14096         -83.96%
BenchmarkAlloc1024Op/PutNew-16          470349        396572        -15.69%
BenchmarkAlloc1024Op/PutReplace-16      321534        247745        -22.95%
BenchmarkAlloc1024Op/SharedTxGet-16     73728         0             -100.00%
BenchmarkDBBatchManual10x100-16         889315        803483        -9.65%
```