@wACKYz 2023-03-23
Ok as I see it is allowed to use multiple goroutines with a single TX read, I will push fix for it.