@wACKYz 2023-03-23
fixed case of multiple goroutines sharing Read TX
benchmarks removed from PR (I can make a separate PR with them if needed)