@serathius 2023-03-24
For performance improvements please use [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat).
You can learn more about it here https://about.sourcegraph.com/blog/go/gophercon-2019-optimizing-go-code-without-a-blindfold