@wACKYz 2023-03-31
```
goos: linux
goarch: amd64
pkg: go.etcd.io/bbolt
cpu: Intel(R) Xeon(R) Gold 6154 CPU @ 3.00GHz
                                  │      org      │              optimized               │
                                  │    sec/op     │    sec/op     vs base                │
DBBatchAutomatic-72                  5.630m ± 14%   5.111m ±  9%   -9.23% (p=0.023 n=10)
DBBatchSingle-72                     242.3m ± 25%   239.6m ± 26%        ~ (p=0.971 n=10)
DB1024OpOnSingleTx/Get-72           2035.1µ ± 10%   825.3µ ± 10%  -59.44% (p=0.000 n=10)
DB1024OpOnSingleTx/PutNew-72         2.385m ± 20%   2.117m ± 43%        ~ (p=0.315 n=10)
DB1024OpOnSingleTx/PutReplace-72     2.142m ± 11%   1.423m ±  7%  -33.59% (p=0.000 n=10)
DB1024OpOnSingleTx/SharedTxGet-72    509.3µ ±  4%   143.4µ ±  2%  -71.84% (p=0.000 n=10)
DBBatchManual10x100-72               5.810m ± 13%   5.548m ± 21%        ~ (p=0.143 n=10)
geomean                              4.575m         3.044m        -33.47%

                                  │     org      │                 optimized                 │
                                  │     B/op     │     B/op      vs base                     │
DBBatchAutomatic-72                 443.0Ki ± 0%   372.1Ki ± 0%   -15.99% (p=0.000 n=10)
DBBatchSingle-72                    17.29Mi ± 5%   17.35Mi ± 5%         ~ (p=0.739 n=10)
DB1024OpOnSingleTx/Get-72           93.27Ki ± 0%   21.10Ki ± 0%   -77.38% (p=0.000 n=10)
DB1024OpOnSingleTx/PutNew-72        467.4Ki ± 0%   395.1Ki ± 0%   -15.47% (p=0.000 n=10)
DB1024OpOnSingleTx/PutReplace-72    321.7Ki ± 0%   249.5Ki ± 0%   -22.46% (p=0.000 n=10)
DB1024OpOnSingleTx/SharedTxGet-72   72.00Ki ± 0%    0.00Ki ± 0%  -100.00% (p=0.000 n=10)
DBBatchManual10x100-72              867.1Ki ± 2%   795.6Ki ± 2%    -8.24% (p=0.000 n=10)
geomean                             490.9Ki                      ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

                                  │     org      │                optimized                 │
                                  │  allocs/op   │  allocs/op   vs base                     │
DBBatchAutomatic-72                 10.194k ± 0%   8.190k ± 0%   -19.66% (p=0.000 n=10)
DBBatchSingle-72                     56.75k ± 0%   54.75k ± 0%    -3.52% (p=0.000 n=10)
DB1024OpOnSingleTx/Get-72           2110.00 ± 0%    62.00 ± 0%   -97.06% (p=0.000 n=10)
DB1024OpOnSingleTx/PutNew-72         3.376k ± 0%   1.327k ± 0%   -60.69% (p=0.000 n=10)
DB1024OpOnSingleTx/PutReplace-72     3.344k ± 0%   1.295k ± 0%   -61.27% (p=0.000 n=10)
DB1024OpOnSingleTx/SharedTxGet-72    2.048k ± 0%   0.000k ± 0%  -100.00% (p=0.000 n=10)
DBBatchManual10x100-72               5.941k ± 1%   3.942k ± 1%   -33.65% (p=0.000 n=10)
geomean                              5.576k                     ?                       ¹ ²
```
The benefit will be visible in large transactions, a single operation/transaction will probably not be noticeable because of  transaction overhead.

Should I include my benchmark in this PR?