@ptabor 2023-04-06
Thank you for the PR. The number looks great. At the same time I'm a little bit afraid of putting changes to the critical logic without growing test coverage. We suspect there is already some subtle memory stomping bug and future changes can mask it even further. I would keep it parked / add to an experimental branch if we had such for now.