@ahrtr 2023-04-06
Thanks @wACKYz . Sorry for the late response. 

This is the most crucial part of the code. I agree with @ptabor , and it's the reason why I do not review it so far. 

I am also thinking to revert the couple of PRs I merged into master branch for now, and get them included in a temporary branch. WDYT? @ptabor @serathius  We need to focus on adding/improving test, the goal is try to reproduce the data corruption issue.